Extract asynchronous processing into separate module
